ADD: Ligação com 'Contrato' e 'Vinculo' em 'Paciente' e 'Representante'

+++ src/main/java/tlk/nexus_core/repositories/PacienteRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import tlk.nexus_core.models.PacienteModel;
 
 @Repository
-public interface PacienteRepository extends JpaRepository<PacienteModel, Integer> {
+public interface PacienteRepository extends JpaRepository<PacienteModel, Long> {
 
   List<PacienteModel> findByNomeContaining(String nome);
